Model relationships added

+++ src/sample/sample/sample.service.ts
@@ -5,6 +5,7 @@ import {SampleRepository} from "./sample.repository";
 import {InjectRepository} from "@nestjs/typeorm";
 import {SampleEntity} from "./sample.entity";
 import {GetSampleFilterDto} from "./dto/getSampleFilter-dto";
+import {User} from "../../auth/user.entity";
 
 @Injectable()
 export class SampleService {
@@ -13,16 +14,16 @@ export class SampleService {
         private sampleRepository: SampleRepository) {
     }
 
-    getSamples(getSampleFilterDto : GetSampleFilterDto) : Promise<SampleEntity[]> {
-        return this.sampleRepository.getSamples(getSampleFilterDto).then(
+    getSamples(getSampleFilterDto : GetSampleFilterDto, user:User) : Promise<SampleEntity[]> {
+        return this.sampleRepository.getSamples(getSampleFilterDto, user).then(
             (samples) => {
                 return samples;
             }
         )
     }
 
-    getSampleById(id:number): Promise<SampleEntity>{
-        return this.sampleRepository.findOne(id).then(
+    getSampleById(id:number, user:User): Promise<SampleEntity>{
+        return this.sampleRepository.findOne({where:{id, userId:user.id}}).then(
             ((found) => {
                 if(!found){
                     throw new BadRequestException(`The sample with ${id} does not exist`);
@@ -34,12 +35,12 @@ export class SampleService {
         )
     }
 
-    createSample(createSampleDto : CreateSampleDto) : Promise<SampleEntity>{
-        return this.sampleRepository.createSample(createSampleDto);
+    createSample(createSampleDto : CreateSampleDto, user:User) : Promise<SampleEntity>{
+        return this.sampleRepository.createSample(createSampleDto, user);
     }
 
-    deleteSample(id: number) {
-        return this.sampleRepository.delete(id).then(
+    deleteSample(id: number, user:User) {
+        return this.sampleRepository.delete({id, userId:user.id}).then(
             (result) => {
                 if(!result.affected){
                     throw new NotFoundException(`Sample with ${id} not found`)
@@ -50,8 +51,8 @@ export class SampleService {
         )
     }
 
-    updateSample(id:number, status: ModelStatus) : Promise<SampleEntity>{
-       return this.getSampleById(id).then(
+    updateSample(id:number, status: ModelStatus, user:User) : Promise<SampleEntity>{
+       return this.getSampleById(id, user).then(
            (sample) => {
                sample.status = status;
                return sample.save().then(
